oviselect-cp-scraper/requirements.txt
requests
scrapling[fetchers]
beautifulsoup4
lxml
pydantic>=2.0
supabase
ollama
tenacity
tqdm
loguru
python-dotenv
thefuzz

